+++ src/orchestration/cli/manual-gates.js
@@ -5,7 +5,7 @@ import { configurationError } from '../../core/error/repo-guard-error.js';
 import { writeGateResultConsole } from '../../core/report/console-renderer.js';
 import { collectProjectFiles } from '../../policies/file-placement.js';
 import { createProjectGateRegistry, gateRegistry } from '../../gates/registry.js';
-import { collectWorkingTreeChanges } from '../../git-changes.js';
+import { collectWorkingTreeChanges } from '../../git/change-collection.js';
 import { findRepositoryRoot } from '../../git.js';
 import { orchestratePlan } from '../orchestrator.js';